[None][refactor] Move Gemma4 fused kernels to _torch/modules/gemma4
_torch/models/ holds only modeling_*.py files; architecture-specific
building blocks belong under _torch/modules/ (like mhc/, engram/,
fused_moe/). Move the four gemma4_fused_* kernel files into a new
modules/gemma4/ subpackage and update imports in modeling_gemma4 and
the unit tests. No functional change.
