+++ tests/unittest/_torch/modules/test_gemma4_fused_gelu_quant.py
@@ -27,7 +27,7 @@ import torch
 
 import tensorrt_llm  # noqa: F401  (registers trtllm torch ops)
 import tensorrt_llm._torch.custom_ops.flashinfer_custom_ops  # noqa: F401
-from tensorrt_llm._torch.models.gemma4_fused_gelu_quant import (
+from tensorrt_llm._torch.modules.gemma4.fused_gelu_quant import (
     gemma4_fused_gelu_mul_fp4,
     sf_swizzled_offsets,
 )
